refactor

+++ app/src/main/java/com/example/angiday/model/entity/UserBehaviorEntity.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "user_behavior")
 data class UserBehaviorEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val userId: Int,
-    val foodId: Int,
+    val userId: Long,
+    val foodId: Long,
     val behaviorType: String, // "favorite", "shared", "cooked"
     val timestamp: String? = null
 )
